requirements.txt
tensorflow==2.19.0
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
joblib==1.5.1
streamlit==1.48.1
Pillow==11.3.0